pyproject.toml
[project]
name = "fastapi-stater"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "booktype>=1.5",
    "certifi==2026.1.4",
    "click==8.3.1",
    "dnspython==2.8.0",
    "email-validator==2.3.0",
    "faker==40.1.2",
    "fastapi==0.128.0",
    "fastapi-cli==0.0.20",
    "fastapi-cloud-cli==0.11.0",
    "fastar==0.8.0",
    "greenlet==3.3.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "idna==3.11",
    "jinja2==3.1.6",
    "loguru>=0.7.3",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "psycopg==3.3.2",
    "psycopg-binary==3.3.2",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-extra-types==2.11.0",
    "pydantic-settings==2.12.0",
    "pygments==2.19.2",
    "pyjwt[cripto,crypto]>=2.12.1",
    "pytest>=9.0.2",
    "python-dotenv==1.2.1",
    "python-multipart==0.0.21",
    "pyyaml==6.0.3",
    "redis[hiredis]>=7.4.0",
    "rich==14.2.0",
    "rich-toolkit==0.17.1",
    "rignore==0.7.6",
    "sentry-sdk==2.49.0",
    "shellingham==1.5.4",
    "slowapi>=0.1.9",
    "sqlalchemy==2.0.45",
    "starlette==0.50.0",
    "testcontainers[postgres]>=4.14.2",
    "typer==0.21.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.3",
    "user-agents>=2.2.0",
    "uvicorn==0.40.0",
    "uvloop==0.22.1",
    "watchfiles==1.1.1",
    "websockets==16.0",
]
